Add files via upload
Identifying webcam images of students in a college, for attendance by linking code with the timetable. First, the student's face is trained and in
real-time testing, it detects the face; details such as date, time, student name, faculty name, and subject
name are taken in a file. The settings are such that the attendance will be marked from a 10 minutes time
interval before and after the class time. Also detects the students with less than 75% attendance.